refactor(typescript): move test helper functionality in a new package
closes #177
