+++ packages/test-helper/CHANGELOG.md
@@ -1,0 +1,11 @@
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+### 0.1.2 (2024-07-11)
+
+### Features
+
+- **test-helper:** add test-helper in runtime libraries repo 
+
