+++ packages/test-helper/jest.config.js
@@ -1,0 +1,7 @@
+const { jest: lernaAliases } = require('lerna-alias');
+
+module.exports = {
+  preset: 'ts-jest',
+  moduleNameMapper: lernaAliases(),
+  coverageReporters: [['lcov', { projectRoot: '../../' }]]
+};
