+++ packages/test-helper/package.json
@@ -1,0 +1,65 @@
+{
+  "name": "@apimatic/test-helper",
+  "version": "0.0.1",
+  "description": "provides the assertion utilities",
+  "main": "lib/index.js",
+  "module": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "engines": {
+    "node": ">=14.15.0 || >=16.0.0"
+  },
+  "scripts": {
+    "clean": "rm -rf lib es umd tsconfig.tsbuildinfo",
+    "test": "jest --passWithNoTests",
+    "build": "npm run clean && tsc && rollup -c && npm run annotate:es",
+    "annotate:es": "babel es --out-dir es --no-babelrc --plugins annotate-pure-calls",
+    "preversion": "npm run test",
+    "prepublishOnly": "npm run build",
+    "size": "size-limit",
+    "analyze": "size-limit --why",
+    "lint": "tslint --project .",
+    "lint:fix": "tslint --project . --fix",
+    "check-style": "prettier --check \"{src,test}/**/*.ts\"",
+    "check-style:fix": "prettier --write \"{src,test}/**/*.ts\""
+  },
+  "author": "APIMatic Ltd.",
+  "license": "ISC",
+  "size-limit": [
+    {
+      "path": "umd/schema.js",
+      "limit": "5 KB"
+    },
+    {
+      "path": "umd/schema.esm.js",
+      "limit": "5 KB"
+    }
+  ],
+  "devDependencies": {
+    "@babel/cli": "^7.21.5",
+    "@babel/core": "^7.22.1",
+    "@rollup/plugin-terser": "^0.4.3",
+    "@size-limit/preset-small-lib": "^7.0.8",
+    "babel-plugin-annotate-pure-calls": "^0.4.0",
+    "jest": "^26.4.2",
+    "jsdom": "^19.0.0",
+    "jsdom-global": "^3.0.2",
+    "lerna-alias": "3.0.3-0",
+    "rollup": "^2.79.0",
+    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-typescript2": "^0.31.0",
+    "size-limit": "^7.0.8",
+    "ts-jest": "^26.4.0",
+    "typescript": "^4.1.2"
+  },
+  "dependencies": {
+    "tslib": "^2.1.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:apimatic/apimatic-js-runtime.git",
+    "directory": "packages/test-helper"
+  }
+}
