+++ packages/test-helper/tsconfig.json
@@ -1,10 +1,10 @@
 {
   "extends": "../../tsconfig.base.json",
-  "include": ["src/**/*", "../test-helper/src/testHelper.ts"],
+  "include": ["src/**/*"],
   "compilerOptions": {
     "outDir": "./lib",
     "rootDir": "./src",
     "skipLibCheck": true,
-    "resolveJsonModule": true
+    "resolveJsonModule": true,
   }
 }
